use java to solve a string question
This is an application that will prompt for a user to input a string. Using the user’s inputted string, find the first letter that is not repeated. 
 
For example: If given the string ‘Bubble’, the letter ‘u’ would be the first character returned from the program. Once the first character is found and displayed back to the user, rewrite the string in order of number of occurrences and order from the inputted string. In the above example ‘Bubble’ would then be rewritten as ‘uleBbb’. Display this to the user.